Move cityhall database to api/db/django
